add tf docker filew

+++ docker_run_gpu.sh
@@ -1,5 +1,17 @@
-docker run --name duckiepond --rm -it --net=host --privileged -v /dev:/dev \
+if [ $# -gt 0 ]; then
+	if [ "$1" == "same" ]; then
+		docker exec -it duckiepond bash
+	else
+		docker run --name duckiepond --rm -it --net=host --privileged -v /dev:/dev \
 			-v /etc/localtime:/etc/localtime:ro -v /var/run/docker.sock:/var/run/docker.sock \
-			-v /home/$USER:/root -v ~/.bashrc:/root/.bashrc \
+			-v /home/$USER/duckiepond-nctu:/root/duckiepond-nctu \
+			-v /home/$USER/duckiepond_gazebo:/root/duckiepond_gazebo \
+			-v /home/$USER/duckiepond-imitaion-learning:/root/duckiepond-imitaion-learning \
 			--runtime=nvidia \
-			tonycar12002/duckiepond:laptop_gpu
+			--env="DISPLAY" --volume="/tmp/.X11-unix:/tmp/.X11-unix:rw" \
+			--env QT_X11_NO_MITSHM=1 \
+			tonycar12002/duckiepond:laptop_gpu     
+	fi
+else
+	echo "please provide docker tag name."
+fi
